forge fmt

+++ src/core/KerosineManager.sol
@@ -2,51 +2,33 @@
 pragma solidity ^0.8.20;
 
 import {EnumerableSet} from "@openzeppelin/contracts/utils/structs/EnumerableSet.sol";
-import {Owned}         from "@solmate/src/auth/Owned.sol";
+import {Owned} from "@solmate/src/auth/Owned.sol";
 
 contract KerosineManager is Owned(msg.sender) {
-  error TooManyVaults();
-  error VaultAlreadyAdded();
-  error VaultNotFound();
+    error TooManyVaults();
+    error VaultAlreadyAdded();
+    error VaultNotFound();
 
-  using EnumerableSet for EnumerableSet.AddressSet;
+    using EnumerableSet for EnumerableSet.AddressSet;
 
-  uint public constant MAX_VAULTS = 10;
+    uint256 public constant MAX_VAULTS = 10;
 
-  EnumerableSet.AddressSet private vaults;
+    EnumerableSet.AddressSet private vaults;
 
-  function add(
-    address vault
-  ) 
-    external 
-      onlyOwner
-  {
-    if (vaults.length() >= MAX_VAULTS) revert TooManyVaults();
-    if (!vaults.add(vault))            revert VaultAlreadyAdded();
-  }
+    function add(address vault) external onlyOwner {
+        if (vaults.length() >= MAX_VAULTS) revert TooManyVaults();
+        if (!vaults.add(vault)) revert VaultAlreadyAdded();
+    }
 
-  function remove(
-    address vault
-  ) 
-    external 
-      onlyOwner
-  {
-    if (!vaults.remove(vault)) revert VaultNotFound();
-  }
+    function remove(address vault) external onlyOwner {
+        if (!vaults.remove(vault)) revert VaultNotFound();
+    }
 
-  function getVaults() 
-    external 
-    view 
-    returns (address[] memory) {
-      return vaults.values();
-  }
+    function getVaults() external view returns (address[] memory) {
+        return vaults.values();
+    }
 
-  function isLicensed(
-    address vault
-  ) 
-    external 
-    view 
-    returns (bool) {
-      return vaults.contains(vault);
-  }
+    function isLicensed(address vault) external view returns (bool) {
+        return vaults.contains(vault);
+    }
 }
