+++ src/staking/Staking.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.8.17;
 
 import {IStaking} from "../interfaces/IStaking.sol";
 
-import {Owned}           from "@solmate/src/auth/Owned.sol";
-import {ERC20}           from "@solmate/src/tokens/ERC20.sol";
+import {Owned} from "@solmate/src/auth/Owned.sol";
+import {ERC20} from "@solmate/src/tokens/ERC20.sol";
 import {SafeTransferLib} from "@solmate/src/utils/SafeTransferLib.sol";
 
 // from https://solidity-by-example.org/defi/staking-rewards/
@@ -15,117 +15,107 @@ contract Staking is IStaking, Owned(msg.sender) {
     ERC20 public immutable rewardsToken;
 
     // Duration of rewards to be paid out (in seconds)
-    uint public duration; 
+    uint256 public duration;
     // Timestamp of when the rewards finish
-    uint public finishAt;
+    uint256 public finishAt;
     // Minimum of last updated time and reward finish time
-    uint public updatedAt;
+    uint256 public updatedAt;
     // Reward to be paid out per second
-    uint public rewardRate;
+    uint256 public rewardRate;
     // Sum of (reward rate * dt * 1e18 / total supply)
-    uint public rewardPerTokenStored;
+    uint256 public rewardPerTokenStored;
     // User address => rewardPerTokenStored
-    mapping(address => uint) public userRewardPerTokenPaid;
+    mapping(address => uint256) public userRewardPerTokenPaid;
     // User address => rewards to be claimed
-    mapping(address => uint) public rewards;
+    mapping(address => uint256) public rewards;
 
     // Total staked
-    uint public totalSupply;
+    uint256 public totalSupply;
     // User address => staked amount
-    mapping(address => uint) public balanceOf;
+    mapping(address => uint256) public balanceOf;
 
     constructor(ERC20 _stakingToken, ERC20 _rewardToken) {
-      stakingToken = _stakingToken;
-      rewardsToken = _rewardToken;
+        stakingToken = _stakingToken;
+        rewardsToken = _rewardToken;
     }
 
     modifier updateReward(address _account) {
-      rewardPerTokenStored = rewardPerToken();
-      updatedAt = lastTimeRewardApplicable();
+        rewardPerTokenStored = rewardPerToken();
+        updatedAt = lastTimeRewardApplicable();
 
-      if (_account != address(0)) {
-          rewards[_account] = earned(_account);
-          userRewardPerTokenPaid[_account] = rewardPerTokenStored;
-      }
+        if (_account != address(0)) {
+            rewards[_account] = earned(_account);
+            userRewardPerTokenPaid[_account] = rewardPerTokenStored;
+        }
 
-      _;
+        _;
     }
 
-    function lastTimeRewardApplicable() public view returns (uint) {
-      return _min(finishAt, block.timestamp);
+    function lastTimeRewardApplicable() public view returns (uint256) {
+        return _min(finishAt, block.timestamp);
     }
 
-    function rewardPerToken() public view returns (uint) {
-      if (totalSupply == 0) {
-          return rewardPerTokenStored;
-      }
+    function rewardPerToken() public view returns (uint256) {
+        if (totalSupply == 0) {
+            return rewardPerTokenStored;
+        }
 
-      return
-          rewardPerTokenStored +
-          (rewardRate * (lastTimeRewardApplicable() - updatedAt) * 1e18) /
-          totalSupply;
+        return rewardPerTokenStored + (rewardRate * (lastTimeRewardApplicable() - updatedAt) * 1e18) / totalSupply;
     }
 
-    function stake(uint _amount) external updateReward(msg.sender) {
-      require(_amount > 0, "amount = 0");
-      stakingToken.safeTransferFrom(msg.sender, address(this), _amount);
-      balanceOf[msg.sender] += _amount;
-      totalSupply += _amount;
-      emit Staked(msg.sender, _amount);
+    function stake(uint256 _amount) external updateReward(msg.sender) {
+        require(_amount > 0, "amount = 0");
+        stakingToken.safeTransferFrom(msg.sender, address(this), _amount);
+        balanceOf[msg.sender] += _amount;
+        totalSupply += _amount;
+        emit Staked(msg.sender, _amount);
     }
 
-    function withdraw(uint _amount) external updateReward(msg.sender) {
-      require(_amount > 0, "amount = 0");
-      balanceOf[msg.sender] -= _amount;
-      totalSupply -= _amount;
-      stakingToken.safeTransfer(msg.sender, _amount);
-      emit Withdrawn(msg.sender, _amount);
+    function withdraw(uint256 _amount) external updateReward(msg.sender) {
+        require(_amount > 0, "amount = 0");
+        balanceOf[msg.sender] -= _amount;
+        totalSupply -= _amount;
+        stakingToken.safeTransfer(msg.sender, _amount);
+        emit Withdrawn(msg.sender, _amount);
     }
 
-    function earned(address _account) public view returns (uint) {
-      return
-          ((balanceOf[_account] *
-              (rewardPerToken() - userRewardPerTokenPaid[_account])) / 1e18) +
-          rewards[_account];
+    function earned(address _account) public view returns (uint256) {
+        return
+            ((balanceOf[_account] * (rewardPerToken() - userRewardPerTokenPaid[_account])) / 1e18) + rewards[_account];
     }
 
     function getReward() external updateReward(msg.sender) {
-      uint reward = rewards[msg.sender];
-      if (reward > 0) {
-          rewards[msg.sender] = 0;
-          rewardsToken.safeTransfer(msg.sender, reward);
-          emit RewardPaid(msg.sender, reward);
-      }
+        uint256 reward = rewards[msg.sender];
+        if (reward > 0) {
+            rewards[msg.sender] = 0;
+            rewardsToken.safeTransfer(msg.sender, reward);
+            emit RewardPaid(msg.sender, reward);
+        }
     }
 
-    function setRewardsDuration(uint _duration) external onlyOwner {
-      require(finishAt < block.timestamp, "reward duration not finished");
-      duration = _duration;
-      emit RewardsDurationUpdated(_duration);
+    function setRewardsDuration(uint256 _duration) external onlyOwner {
+        require(finishAt < block.timestamp, "reward duration not finished");
+        duration = _duration;
+        emit RewardsDurationUpdated(_duration);
     }
 
-    function notifyRewardAmount(
-      uint _amount
-    ) external onlyOwner updateReward(address(0)) {
-      if (block.timestamp >= finishAt) {
-          rewardRate = _amount / duration;
-      } else {
-          uint remainingRewards = (finishAt - block.timestamp) * rewardRate;
-          rewardRate = (_amount + remainingRewards) / duration;
-      }
+    function notifyRewardAmount(uint256 _amount) external onlyOwner updateReward(address(0)) {
+        if (block.timestamp >= finishAt) {
+            rewardRate = _amount / duration;
+        } else {
+            uint256 remainingRewards = (finishAt - block.timestamp) * rewardRate;
+            rewardRate = (_amount + remainingRewards) / duration;
+        }
 
-      require(rewardRate > 0, "reward rate = 0");
-      require(
-          rewardRate * duration <= rewardsToken.balanceOf(address(this)),
-          "reward amount > balance"
-      );
+        require(rewardRate > 0, "reward rate = 0");
+        require(rewardRate * duration <= rewardsToken.balanceOf(address(this)), "reward amount > balance");
 
-      finishAt = block.timestamp + duration;
-      updatedAt = block.timestamp;
-      emit RewardAdded(_amount);
+        finishAt = block.timestamp + duration;
+        updatedAt = block.timestamp;
+        emit RewardAdded(_amount);
     }
 
-    function _min(uint x, uint y) private pure returns (uint) {
-      return x <= y ? x : y;
+    function _min(uint256 x, uint256 y) private pure returns (uint256) {
+        return x <= y ? x : y;
     }
 }
